fix(wip): Get most of the way through the OIDC plugin migration
