+++ cmd/account.go
@@ -4,23 +4,16 @@ import (
 	"errors"
 	"fmt"
 	"log"
-	"slices"
-	"strings"
 
+	lacct "github.com/louislef299/aws-sso/internal/account"
 	lenv "github.com/louislef299/aws-sso/internal/envs"
 	lregion "github.com/louislef299/aws-sso/internal/region"
-	laws "github.com/louislef299/aws-sso/pkg/aws"
-	los "github.com/louislef299/aws-sso/pkg/os"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 )
 
 const (
-	ACCOUNT_GROUP  = "account"
-	ACCOUNT_REGION = "region"
-	ACCOUNT_ID     = "id"
-	ACCOUNT_URL    = "url"
-	ROLE_ARN_KEY   = "role_arn"
+	ROLE_ARN_KEY = "role_arn"
 )
 
 var (
@@ -35,14 +28,6 @@ var (
 	ErrNoAccountFound = errors.New("no account found")
 	ErrAccountsToFix  = errors.New("accounts need to be fixed")
 )
-
-type Account struct {
-	ID      string
-	Private bool
-	Region  string
-	Token   string
-	URL     string
-}
 
 // accountCmd represents the account command
 var accountCmd = &cobra.Command{
@@ -68,10 +53,10 @@ var accountAddCmd = &cobra.Command{
 			}
 		}
 		if accountURL == "" {
-			accountURL = getAccountURL(args[0])
+			accountURL = lacct.GetAccountURL(args[0])
 		}
 
-		err = addAccount(accountName, &Account{
+		err = lacct.AddAccount(accountName, &lacct.Account{
 			ID:      accountNumber,
 			Region:  accountRegion,
 			Private: accountPrivate,
@@ -93,7 +78,7 @@ var accountListCmd = &cobra.Command{
 	Long: `Lists the local AWS account alias mapping and AWS name
 profiles found.`,
 	Run: func(cmd *cobra.Command, args []string) {
-		listAccounts(allAccounts)
+		lacct.ListAccounts(allAccounts)
 	},
 }
 
@@ -105,16 +90,16 @@ var accountSetCmd = &cobra.Command{
 	Example: `  aws-sso account set env1 --number 000000000 --region us-west-2 --token dev --private`,
 	Run: func(cmd *cobra.Command, args []string) {
 		if accountNumber == "" {
-			accountNumber = getAccountID(args[0])
+			accountNumber = lacct.GetAccountID(args[0])
 		}
 		if accountRegion == "" {
-			accountRegion = getAccountRegion(args[0])
+			accountRegion = lacct.GetAccountRegion(args[0])
 		}
 		if accountURL == "" {
-			accountURL = getAccountURL(args[0])
+			accountURL = lacct.GetAccountURL(args[0])
 		}
 
-		viperSetAccount(args[0], Account{
+		lacct.ViperSetAccount(args[0], lacct.Account{
 			ID:      accountNumber,
 			Region:  accountRegion,
 			Private: accountPrivate,
@@ -136,7 +121,7 @@ var accountPluralCmd = &cobra.Command{
 	Aliases: []string{"accts"},
 	Hidden:  true,
 	Run: func(cmd *cobra.Command, args []string) {
-		listAccounts(allAccounts)
+		lacct.ListAccounts(allAccounts)
 	},
 }
 
@@ -171,125 +156,11 @@ func init() {
 	accountSetCmd.Flags().BoolVarP(&accountPrivate, "private", "p", false, "The account is a private account")
 }
 
-func addAccount(name string, account *Account) error {
-	viperSetAccount(name, *account)
-	return viper.WriteConfig()
-}
-
-// Sets account value in Viper. Does not write to config
-func viperSetAccount(name string, acct Account) {
-	viper.Set(fmt.Sprintf("account.%s", name), acct)
-}
-
-func getAccountID(profile string) string {
-	id := viper.GetString(fmt.Sprintf("%s.%s.%s", ACCOUNT_GROUP, profile, ACCOUNT_ID))
-	if id == "" {
-		log.Printf("couldn't find an account ID matching profile %s, using empty default...\n", profile)
-	}
-	return id
-}
-
-func getAccountURL(profile string) string {
-	url := viper.GetString(fmt.Sprintf("%s.%s.%s", ACCOUNT_GROUP, profile, ACCOUNT_URL))
-	if url == "" {
-		log.Printf("couldn't find an account URL matching profile %s, using core default...\n", url)
-	} else {
-		return url
-	}
-	return viper.GetString(lenv.CORE_URL)
-}
-
-func getAccountRegion(profile string) string {
-	var err error
-	r := viper.GetString(fmt.Sprintf("%s.%s.%s", ACCOUNT_GROUP, profile, ACCOUNT_REGION))
-	if r == "" {
-		log.Printf("couldn't find an account region matching profile %s, using local default...\n", profile)
-		r, err = lregion.GetRegion(lregion.EKS)
-		if err != nil {
-			log.Fatal("couldn't get default region:", err)
-		}
-	}
-	return r
-}
-
-func getAccountToken(profile string) string {
-	t := viper.GetString(fmt.Sprintf("%s.%s.%s", ACCOUNT_GROUP, profile, "token"))
-	if t == "" {
-		return "default"
-	}
-	return t
-}
-
-func getAccountPrivate(profile string) bool {
-	return viper.GetBool(fmt.Sprintf("%s.%s.%s", ACCOUNT_GROUP, profile, "private"))
-}
-
-func listAccounts(all bool) {
-	accts := viper.Sub(ACCOUNT_GROUP)
-	if accts == nil {
-		log.Println("no accounts configured! use the 'account add' command to create new mappings")
-		return
-	}
-
-	fmt.Println("Account mapping:")
-	acctKeys := accts.AllKeys()
-	var acctList []string
-	for _, k := range acctKeys {
-		acctList = append(acctList, trimSuffixes(k))
-	}
-	slices.Sort(acctList)
-	acctList = slices.Compact(acctList)
-
-	for _, a := range acctList {
-		var account Account
-		err := accts.UnmarshalKey(a, &account)
-		var t string
-		if account.Token == "" {
-			t = "default"
-		} else {
-			t = account.Token
-		}
-
-		var url string
-		if account.URL == "" {
-			url = "(default) " + viper.GetString(lenv.CORE_URL)
-		} else {
-			url = account.URL
-		}
-
-		if err == nil {
-			fmt.Printf("%s:\n  ID: %s\n  Region: %s\n  Private: %t\n  Token: %s\n  SSO URL: %s\n",
-				a, account.ID, account.Region, account.Private, t, url)
-		}
-	}
-
-	if all {
-		fmt.Printf("\nAWS Configs:\n")
-		sections, err := laws.GetAWSProfiles()
-		if err != nil {
-			log.Fatal(err)
-		}
-		for _, s := range sections {
-			if !los.IsManagedProfile(s) {
-				fmt.Println(s)
-			}
-		}
-	}
-}
-
 func syncAccountRegionSession(profile, region string) (string, error) {
 	if region == "" {
-		region = getAccountRegion(profile)
+		region = lacct.GetAccountRegion(profile)
 	}
 	viper.Set(lenv.SESSION_REGION, region)
 	err := viper.WriteConfig()
 	return region, err
-}
-
-func trimSuffixes(s string) string {
-	suffixes := []string{"id", "region"}
-	for _, suff := range suffixes {
-		s = strings.TrimSuffix(s, fmt.Sprintf(".%s", suff))
-	}
-	return s
 }
